Improved handling of setting/removing card animations
Removed animation end event code as it is not fully supported. Replaced
it with code that changes a dependency for when a card has been
randomly selected, then each card checks to see if it has been
animated. If it has (it has the class .animated), then a timeout is set
to remove the class, thus completing the effect. These dependencies and
timeouts are cleared when the card templates are destroyed.