Project background images now loading.
Different background images now specified in CSS. We need to do this
because of Meteors idiosyncratic way of handling template data.
Rendered, created and destroyed sometimes have the full project data
and sometimes not. The routes are seemingly unable to fetch the data,
waiting until its all there before rendering the template.